+++ packages/nuxt-locokit/src/runtime/plugins/1_error.ts
