+++ packages/nuxt-locokit/src/runtime/plugins/2_directive.ts
