+++ packages/nuxt-locokit/src/runtime/plugins/3_i18n.ts
