+++ packages/nuxt-locokit/src/runtime/plugins/4_primevue.ts
