+++ packages/nuxt-locokit/src/runtime/plugins/5_vee-validate.ts
@@ -1,7 +1,7 @@
 // import { isAfter, isValid } from 'date-fns'
-import { defineNuxtPlugin } from '#app'
 import { setupVeeValidate } from '@locokit/designsystem'
-import { i18n } from './i18n'
+import { i18n } from './3_i18n'
+import { defineNuxtPlugin } from '#app'
 
 export default defineNuxtPlugin(() => {
   setupVeeValidate(i18n)
